♻️ extract segments
